@@ -1,7 +1,7 @@
 // Top-level build file where you can add configuration options common to all sub-projects/modules.
 
 buildscript {
-    ext.kotlin_version = '1.2.51'
+    ext.kotlin_version = '1.3.11'
     repositories {
         google()
         jcenter()
@@ -10,13 +10,13 @@ buildscript {
         }
     }
     dependencies {
-        classpath 'com.android.tools.build:gradle:3.3.0-alpha10'
+        classpath 'com.android.tools.build:gradle:3.4.0-rc02'
         classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version"
         classpath "org.jetbrains.kotlin:kotlin-allopen:$kotlin_version"
         classpath 'com.google.gms:google-services:4.0.1'
-        classpath 'io.fabric.tools:gradle:1.25.4'
+        classpath 'io.fabric.tools:gradle:1.26.0'
         classpath 'com.google.firebase:firebase-plugins:1.1.5'
-        classpath 'android.arch.navigation:navigation-safe-args-gradle-plugin:1.0.0-alpha04'
+        classpath 'android.arch.navigation:navigation-safe-args-gradle-plugin:1.0.0-alpha07'
 
     }
 }
@@ -30,7 +30,10 @@ allprojects {
     repositories {
         google()
         jcenter()
+        maven { url "https://jitpack.io" }
+
     }
+    apply from: "$rootDir/ktlint.gradle"
 }
 
 repositories {
