@@ -2,7 +2,7 @@ package cat.xlagunas.viv.commons.di
 
 import android.app.Application
 import android.content.Context
-import android.support.test.runner.AndroidJUnitRunner
+import androidx.test.runner.AndroidJUnitRunner
 import cat.xlagunas.viv.commons.TestApplication
 
 class MockTestRunner : AndroidJUnitRunner() {
