@@ -1,24 +1,10 @@
 package cat.xlagunas.viv.commons
 
-import android.app.Application
-import android.app.Service
-import cat.xlagunas.viv.commons.di.DaggerApplicationTestComponent
-import dagger.android.AndroidInjector
-import dagger.android.DispatchingAndroidInjector
-import dagger.android.HasServiceInjector
-import javax.inject.Inject
+import cat.xlagunas.core.di.VivApplication
 
-class TestApplication : Application(), HasServiceInjector {
-
-    @Inject
-    lateinit var dispatchServiceInject: DispatchingAndroidInjector<Service>
+class TestApplication : VivApplication() {
 
     override fun onCreate() {
         super.onCreate()
-        DaggerApplicationTestComponent.builder().withApplication(this).build().inject(this)
-    }
-
-    override fun serviceInjector(): AndroidInjector<Service> {
-        return dispatchServiceInject
     }
 }
