@@ -1,34 +1,32 @@
 package cat.xlagunas.viv.contact
 
-import android.arch.core.executor.testing.InstantTaskExecutorRule
-import android.arch.lifecycle.MutableLiveData
-import android.support.test.espresso.Espresso.onView
-import android.support.test.espresso.UiController
-import android.support.test.espresso.ViewAction
-import android.support.test.espresso.assertion.ViewAssertions
-import android.support.test.espresso.contrib.RecyclerViewActions
-import android.support.test.espresso.matcher.ViewMatchers
-import android.support.test.espresso.matcher.ViewMatchers.withId
-import android.support.test.espresso.matcher.ViewMatchers.withText
-import android.support.test.rule.ActivityTestRule
-import android.support.test.runner.AndroidJUnit4
-import android.support.v7.widget.RecyclerView
+import android.os.Bundle
 import android.view.View
+import androidx.fragment.app.Fragment
+import androidx.fragment.app.FragmentFactory
+import androidx.fragment.app.testing.launchFragmentInContainer
+import androidx.lifecycle.MutableLiveData
+import androidx.lifecycle.ViewModelProvider
 import androidx.navigation.NavController
-import cat.xlagunas.data.common.net.Relationship
-import cat.xlagunas.domain.commons.Friend
+import androidx.test.espresso.Espresso.onView
+import androidx.test.espresso.UiController
+import androidx.test.espresso.ViewAction
+import androidx.test.espresso.assertion.ViewAssertions
+import androidx.test.espresso.contrib.RecyclerViewActions
+import androidx.test.espresso.matcher.ViewMatchers
+import androidx.test.espresso.matcher.ViewMatchers.withId
+import androidx.test.espresso.matcher.ViewMatchers.withText
+import androidx.test.runner.AndroidJUnit4
+import cat.xlagunas.core.domain.entity.Friend
 import cat.xlagunas.viv.R
 import cat.xlagunas.viv.commons.ViewModelUtil
 import cat.xlagunas.viv.contact.viewholder.ConfirmFriendViewHolder
 import cat.xlagunas.viv.contact.viewholder.CurrentFriendViewHolder
 import cat.xlagunas.viv.contact.viewholder.RequestFriendViewHolder
 import cat.xlagunas.viv.push.PushTokenPresenter
-import cat.xlagunas.viv.test.SingleFragmentActivity
 import org.hamcrest.Matcher
 import org.junit.Before
-import org.junit.Rule
 import org.junit.Test
-import org.junit.rules.TestRule
 import org.junit.runner.RunWith
 import org.mockito.Mockito
 import org.mockito.Mockito.`when`
@@ -38,34 +36,20 @@ import org.mockito.Mockito.verify
 @RunWith(AndroidJUnit4::class)
 class ContactFragmentTest {
 
-    @Rule
-    @JvmField
-    val activityRule = ActivityTestRule(SingleFragmentActivity::class.java, true, true)
-
-    @Rule
-    @JvmField
-    var rule: TestRule = InstantTaskExecutorRule()
-
-    private val testFragment = TestContactFragment()
-    private lateinit var contactViewModel: ContactViewModel
-    private lateinit var pushTokenPresenter: PushTokenPresenter
-
+    private val contactViewModel = mock(ContactViewModel::class.java)
+    private val pushTokenPresenter = mock(PushTokenPresenter::class.java)
     private val contactsLiveData = MutableLiveData<List<Friend>>()
+
+    private lateinit var testContactFactory: TestContactFragmentFactory
 
     @Before
     fun setUp() {
-        contactViewModel = mock(ContactViewModel::class.java)
-        pushTokenPresenter = mock(PushTokenPresenter::class.java)
-
-        val contactViewModelFactory = ViewModelUtil.createFor(contactViewModel)
-
-        testFragment.viewModelFactory = contactViewModelFactory
-        testFragment.pushTokenPresenter = pushTokenPresenter
-
-        activityRule.activity.setFragment(testFragment)
+        testContactFactory = TestContactFragmentFactory(ViewModelUtil.createFor(contactViewModel), pushTokenPresenter)
 
         `when`(pushTokenPresenter.isPushTokenRegistered()).thenReturn(false)
         `when`(contactViewModel.contacts).thenReturn(contactsLiveData)
+
+        launchFragmentInContainer<TestContactFragment>(Bundle(), R.style.AppTheme_NoActionBar, testContactFactory)
     }
 
     @Test
@@ -81,35 +65,68 @@ class ContactFragmentTest {
 
     @Test
     fun givenRequestedContact_whenAcceptingContact_thenAcceptCall() {
-        val friend = Friend(0, "FirstFriend", "First friend name", null, "first@gmail.com", Relationship.PENDING.name)
+        val friend = Friend(
+            0,
+            "FirstFriend",
+            "First friend name",
+            null,
+            "first@gmail.com",
+            cat.xlagunas.core.data.net.Relationship.PENDING.name
+        )
 
-        setupFriendRelationshipAndViewHolderType<ConfirmFriendViewHolder>(friend, R.id.accept_friendship_button, 0)
+        setupFriendRelationshipAndViewHolderType<ConfirmFriendViewHolder>(
+            friend,
+            R.id.accept_friendship_button,
+            0
+        )
 
         verify(contactViewModel).acceptContactRequest(friend)
     }
 
     @Test
     fun givenRequestedContact_whenRejectedContact_thenRejectCall() {
-        val friend = Friend(0, "FirstFriend", "First friend name", null, "first@gmail.com", Relationship.PENDING.name)
-        setupFriendRelationshipAndViewHolderType<ConfirmFriendViewHolder>(friend, R.id.reject_friendship_button, 0)
+        val friend = Friend(
+            0,
+            "FirstFriend",
+            "First friend name",
+            null,
+            "first@gmail.com",
+            cat.xlagunas.core.data.net.Relationship.PENDING.name
+        )
+        setupFriendRelationshipAndViewHolderType<ConfirmFriendViewHolder>(
+            friend,
+            R.id.reject_friendship_button,
+            0
+        )
 
         verify(contactViewModel).rejectContactRequest(friend)
     }
 
     @Test
     fun givenAcceptedContact_whenClicked_thenCall() {
-        val friend = Friend(0, "FirstFriend", "First friend name", null, "first@gmail.com", Relationship.ACCEPTED.name)
-        setupFriendRelationshipAndViewHolderType<CurrentFriendViewHolder>(friend, R.id.call_friend_button, 0)
+        val friend = Friend(
+            0,
+            "FirstFriend",
+            "First friend name",
+            null,
+            "first@gmail.com",
+            cat.xlagunas.core.data.net.Relationship.ACCEPTED.name
+        )
+        setupFriendRelationshipAndViewHolderType<CurrentFriendViewHolder>(
+            friend,
+            R.id.call_friend_button,
+            0
+        )
 
         verify(contactViewModel).callFriend(friend)
     }
 
-    private fun <VH : RecyclerView.ViewHolder> setupFriendRelationshipAndViewHolderType(
+    private fun <VH : androidx.recyclerview.widget.RecyclerView.ViewHolder> setupFriendRelationshipAndViewHolderType(
         friend: Friend,
         actionId: Int,
         holderPosition: Int
     ) {
-        contactsLiveData.value = listOf(friend)
+        contactsLiveData.postValue(listOf(friend))
 
         onView(withId(R.id.recycler_view))
             .perform(
@@ -120,7 +137,14 @@ class ContactFragmentTest {
     }
 
     private fun populateContactsList(): List<Friend> {
-        val friend = Friend(0, "FirstFriend", "First friend name", null, "first@gmail.com", Relationship.REQUESTED.name)
+        val friend = Friend(
+            0,
+            "FirstFriend",
+            "First friend name",
+            null,
+            "first@gmail.com",
+            cat.xlagunas.core.data.net.Relationship.REQUESTED.name
+        )
         val friendList = ArrayList<Friend>(1000)
         friendList += friend
 
@@ -138,6 +162,25 @@ class ContactFragmentTest {
     class TestContactFragment : ContactFragment() {
         val navController = Mockito.mock(NavController::class.java)
         override fun navController() = navController
+
+        override fun inject() {}
+    }
+
+    class TestContactFragmentFactory(
+        private val contactViewModelFactory: ViewModelProvider.Factory,
+        private val pushPresenter: PushTokenPresenter
+    ) : FragmentFactory() {
+        override fun instantiate(
+            classLoader: ClassLoader,
+            className: String,
+            args: Bundle?
+        ): Fragment {
+            return TestContactFragment()
+                .apply {
+                    viewModelFactory = contactViewModelFactory
+                    pushTokenPresenter = pushPresenter
+                }
+        }
     }
 
     object MyViewAction {
