@@ -1,6 +1,6 @@
 package cat.xlagunas.viv.commons
 
-import android.arch.lifecycle.ViewModel
+import androidx.lifecycle.ViewModel
 import io.reactivex.disposables.CompositeDisposable
 
 abstract class DisposableViewModel : ViewModel() {
