@@ -1,6 +1,6 @@
 package cat.xlagunas.viv.commons.binding
 
-import android.databinding.BaseObservable
+import androidx.databinding.BaseObservable
 import kotlin.reflect.KProperty
 
 class BindableDelegate<in R : BaseObservable, T : Any>(private var value: T, private val bindingEntry: Int) {
