@@ -1,10 +1,8 @@
 package cat.xlagunas.viv.contact
 
-import android.support.v7.widget.RecyclerView
 import android.view.LayoutInflater
 import android.view.ViewGroup
-import cat.xlagunas.data.common.net.Relationship
-import cat.xlagunas.domain.commons.Friend
+import cat.xlagunas.core.domain.entity.Friend
 import cat.xlagunas.viv.R
 import cat.xlagunas.viv.contact.viewholder.ConfirmFriendViewHolder
 import cat.xlagunas.viv.contact.viewholder.CurrentFriendViewHolder
@@ -13,11 +11,14 @@ import cat.xlagunas.viv.contact.viewholder.PendingFriendViewHolder
 import cat.xlagunas.viv.contact.viewholder.RequestFriendViewHolder
 
 class ContactAdapter constructor(private val contactListener: ContactListener) :
-    RecyclerView.Adapter<RecyclerView.ViewHolder>() {
+    androidx.recyclerview.widget.RecyclerView.Adapter<androidx.recyclerview.widget.RecyclerView.ViewHolder>() {
 
     private var items = emptyList<Friend>()
 
-    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): RecyclerView.ViewHolder {
+    override fun onCreateViewHolder(
+        parent: ViewGroup,
+        viewType: Int
+    ): androidx.recyclerview.widget.RecyclerView.ViewHolder {
         val holder = LayoutInflater.from(parent.context).inflate(viewType, parent, false)
         return when (viewType) {
             R.layout.row_request_contact -> RequestFriendViewHolder(holder, contactListener)
@@ -30,7 +31,10 @@ class ContactAdapter constructor(private val contactListener: ContactListener) :
 
     override fun getItemCount() = items.size
 
-    override fun onBindViewHolder(holder: RecyclerView.ViewHolder, position: Int) {
+    override fun onBindViewHolder(
+        holder: androidx.recyclerview.widget.RecyclerView.ViewHolder,
+        position: Int
+    ) {
         (holder as FriendViewHolder).bind(items[position])
     }
 
@@ -41,10 +45,10 @@ class ContactAdapter constructor(private val contactListener: ContactListener) :
 
     override fun getItemViewType(position: Int): Int {
         return when (items[position].relationshipStatus) {
-            Relationship.NONE.name -> R.layout.row_request_contact
-            Relationship.REQUESTED.name -> R.layout.row_pending_contact
-            Relationship.ACCEPTED.name -> R.layout.row_contact
-            Relationship.PENDING.name -> R.layout.row_confirm_contact
+            cat.xlagunas.core.data.net.Relationship.NONE.name -> R.layout.row_request_contact
+            cat.xlagunas.core.data.net.Relationship.REQUESTED.name -> R.layout.row_pending_contact
+            cat.xlagunas.core.data.net.Relationship.ACCEPTED.name -> R.layout.row_contact
+            cat.xlagunas.core.data.net.Relationship.PENDING.name -> R.layout.row_confirm_contact
             else -> throw IllegalStateException("This should never be called")
         }
     }
