@@ -1,17 +1,17 @@
 package cat.xlagunas.viv.login
 
-import cat.xlagunas.data.common.provider.ActivityMonitor
+import cat.xlagunas.core.domain.schedulers.RxSchedulers
 import cat.xlagunas.data.user.login.GoogleSignInDataSource
-import cat.xlagunas.domain.schedulers.RxSchedulers
 import dagger.Module
 import dagger.Provides
-import javax.inject.Singleton
 
 @Module
 class LoginModule {
 
     @Provides
-    @Singleton
-    fun provideGoogleSignIn(activityMonitor: ActivityMonitor, rxSchedulers: RxSchedulers) =
+    fun provideGoogleSignIn(
+        activityMonitor: cat.xlagunas.core.data.provider.ActivityMonitor,
+        rxSchedulers: RxSchedulers
+    ) =
         GoogleSignInDataSource(activityMonitor, rxSchedulers)
 }
