@@ -1,9 +1,9 @@
 package cat.xlagunas.viv.login
 
-import android.arch.lifecycle.LifecycleObserver
-import android.arch.lifecycle.LiveData
-import android.arch.lifecycle.MutableLiveData
-import android.arch.lifecycle.ViewModel
+import androidx.lifecycle.LifecycleObserver
+import androidx.lifecycle.LiveData
+import androidx.lifecycle.MutableLiveData
+import androidx.lifecycle.ViewModel
 import cat.xlagunas.data.OpenForTesting
 import cat.xlagunas.data.user.login.GoogleSignInDataSource
 import cat.xlagunas.domain.user.authentication.AuthenticationCredentials
