@@ -1,41 +1,52 @@
 package cat.xlagunas.viv.register
 
-import android.arch.lifecycle.Observer
-import android.arch.lifecycle.ViewModelProvider
-import android.arch.lifecycle.ViewModelProviders
-import android.databinding.DataBindingUtil
 import android.os.Bundle
-import android.support.design.widget.FloatingActionButton
-import android.support.design.widget.Snackbar
-import android.support.v4.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import androidx.databinding.DataBindingUtil
+import androidx.lifecycle.Observer
+import androidx.lifecycle.ViewModelProvider
+import androidx.lifecycle.ViewModelProviders
 import androidx.navigation.fragment.findNavController
+import cat.xlagunas.core.di.Injectable
+import cat.xlagunas.core.di.VivApplication
+import cat.xlagunas.core.domain.entity.User
 import cat.xlagunas.data.OpenForTesting
-import cat.xlagunas.domain.commons.User
 import cat.xlagunas.viv.R
-import cat.xlagunas.viv.commons.di.Injectable
 import cat.xlagunas.viv.databinding.ActivityRegisterBinding
+import com.google.android.material.snackbar.Snackbar
+import com.google.android.material.snackbar.Snackbar.LENGTH_LONG
+import dagger.DaggerMonolythComponent
 import timber.log.Timber
 import javax.inject.Inject
 
 @OpenForTesting
-class RegisterFragment : Fragment(), Injectable {
+class RegisterFragment : androidx.fragment.app.Fragment(), Injectable {
 
     private lateinit var registerViewModel: RegisterViewModel
-    private lateinit var actionButton: FloatingActionButton
+    private lateinit var actionButton: com.google.android.material.floatingactionbutton.FloatingActionButton
     private lateinit var binding: ActivityRegisterBinding
 
     @Inject
     lateinit var viewModelFactory: ViewModelProvider.Factory
 
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        inject()
+    }
+
+    protected fun inject() {
+        DaggerMonolythComponent.builder()
+            .withParentComponent(VivApplication.appComponent(context!!)).build()
+            .inject(this)
+    }
+
     override fun onActivityCreated(savedInstanceState: Bundle?) {
         super.onActivityCreated(savedInstanceState)
         registerViewModel = ViewModelProviders.of(this, viewModelFactory).get(RegisterViewModel::class.java)
-        registerViewModel.findUser().observe(this, Observer<User> {
-            Timber.d("User logged in %s", it?.username)
-        })
+        registerViewModel.findUser().observe(this, Observer(this::handleUserAlreadyLoggedIn))
+        registerViewModel.registrationState.observe(this, Observer(this::handleRegistration))
     }
 
     override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? {
@@ -48,27 +59,31 @@ class RegisterFragment : Fragment(), Injectable {
         super.onViewCreated(view, savedInstanceState)
         actionButton = view.findViewById(R.id.fab)
         actionButton.setOnClickListener {
-            registerViewModel
-                .register(RegisterUserConverter().toUser(binding.user!!))
-                .doOnSubscribe { actionButton.isEnabled = false }
-                .subscribe({
-                    showToast()
-                    actionButton.isEnabled = true
-                    navController().navigate(R.id.action_register_successful)
-                }, { error ->
-                    Timber.e(error, "Network error")
-                    actionButton.isEnabled = true
-                    showErrorToast()
-                })
+            registerViewModel.register(RegisterUserConverter().toUser(binding.user!!))
         }
     }
 
-    private fun showToast() {
-        Snackbar.make(binding.root, "User registered ", Snackbar.LENGTH_LONG).show()
+    private fun handleNavigationSuccess() {
+        Snackbar.make(binding.root, "User registered ", LENGTH_LONG).show()
+        navController().navigate(R.id.action_register_successful)
     }
 
-    private fun showErrorToast() {
-        Snackbar.make(binding.root, "Error registering user", Snackbar.LENGTH_LONG).show()
+    private fun handleUserAlreadyLoggedIn(user: User?) {
+        Timber.d("User logged in ${user?.username}")
+    }
+
+    private fun handleRegistration(registrationState: RegistrationState?) {
+        when (registrationState) {
+            is RegistrationError -> handleError(registrationState.message)
+            is Success -> handleNavigationSuccess()
+        }
+        actionButton.isEnabled = true
+    }
+
+    private fun handleError(message: String?) {
+        val errorMsg = "Error registering user $message"
+        Timber.e(errorMsg)
+        Snackbar.make(binding.root, errorMsg, LENGTH_LONG).show()
     }
 
     fun navController() = findNavController()
