@@ -1,12 +1,11 @@
 package cat.xlagunas.data.user.login
 
 import android.app.Activity
-import android.arch.lifecycle.Lifecycle
-import android.arch.lifecycle.LifecycleObserver
-import android.arch.lifecycle.OnLifecycleEvent
-import cat.xlagunas.data.common.provider.ActivityMonitor
-import cat.xlagunas.domain.commons.User
-import cat.xlagunas.domain.schedulers.RxSchedulers
+import androidx.lifecycle.Lifecycle
+import androidx.lifecycle.LifecycleObserver
+import androidx.lifecycle.OnLifecycleEvent
+import cat.xlagunas.core.domain.entity.User
+import cat.xlagunas.core.domain.schedulers.RxSchedulers
 import com.google.android.gms.auth.api.signin.GoogleSignIn
 import com.google.android.gms.auth.api.signin.GoogleSignInAccount
 import com.google.android.gms.auth.api.signin.GoogleSignInClient
@@ -19,7 +18,7 @@ import java.util.UUID
 import javax.inject.Inject
 
 class GoogleSignInDataSource @Inject constructor(
-    private val activityMonitor: ActivityMonitor,
+    private val activityMonitor: cat.xlagunas.core.data.provider.ActivityMonitor,
     private val schedulers: RxSchedulers
 ) : LifecycleObserver {
 
@@ -67,15 +66,20 @@ class GoogleSignInDataSource @Inject constructor(
     fun handleSignInResult(completedTask: Task<GoogleSignInAccount>): Flowable<User> {
         return try {
             val account = completedTask.getResult(ApiException::class.java)
-            val user = User(
-                account.email!!,
-                account.givenName!!,
-                account.familyName!!,
-                account.email!!,
-                account.photoUrl.toString(),
-                UUID.randomUUID().toString()
-            )
-            Flowable.just(user)
+            return if (account != null) {
+                Flowable.just(
+                    User(
+                        account.email!!,
+                        account.givenName!!,
+                        account.familyName!!,
+                        account.email!!,
+                        account.photoUrl.toString(),
+                        UUID.randomUUID().toString()
+                    )
+                )
+            } else {
+                return Flowable.empty()
+            }
         } catch (e: ApiException) {
             Flowable.just(User("This", "was", "a", "user", "who", "didn't"))
         }
