@@ -1,15 +1,8 @@
 package cat.xlagunas.data.contact
 
-import android.arch.core.executor.testing.InstantTaskExecutorRule
-import android.arch.persistence.room.Room
-import cat.xlagunas.data.BuildConfig
-import cat.xlagunas.data.common.converter.FriendConverter
-import cat.xlagunas.data.common.db.FriendDao
-import cat.xlagunas.data.common.db.UserDao
-import cat.xlagunas.data.common.db.UserEntity
-import cat.xlagunas.data.common.db.VivDatabase
-import cat.xlagunas.data.common.net.Relationship
-import cat.xlagunas.domain.commons.Friend
+import androidx.arch.core.executor.testing.InstantTaskExecutorRule
+import androidx.room.Room
+import cat.xlagunas.core.domain.entity.Friend
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.After
 import org.junit.Before
@@ -19,23 +12,21 @@ import org.junit.runner.RunWith
 import org.mockito.MockitoAnnotations
 import org.robolectric.RobolectricTestRunner
 import org.robolectric.RuntimeEnvironment
-import org.robolectric.annotation.Config
 
 @RunWith(RobolectricTestRunner::class)
-@Config(constants = BuildConfig::class, sdk = [26])
 class LocalContactDataSourceTest {
 
     private lateinit var localContactDataSource: LocalContactDataSource
 
-    private lateinit var userDao: UserDao
+    private lateinit var userDao: cat.xlagunas.core.data.db.UserDao
 
-    private lateinit var friendDao: FriendDao
+    private lateinit var friendDao: cat.xlagunas.core.data.db.FriendDao
 
-    private lateinit var vivDatabase: VivDatabase
+    private lateinit var vivDatabase: cat.xlagunas.core.data.db.VivDatabase
 
-    private lateinit var userEntity: UserEntity
+    private lateinit var userEntity: cat.xlagunas.core.data.db.UserEntity
 
-    private val friendConverter = FriendConverter()
+    private val friendConverter = cat.xlagunas.core.data.converter.FriendConverter()
 
     @get:Rule
     var instantTaskExecutorRule = InstantTaskExecutorRule()
@@ -43,11 +34,11 @@ class LocalContactDataSourceTest {
     @Before
     fun setUp() {
         MockitoAnnotations.initMocks(this)
-        vivDatabase = Room.inMemoryDatabaseBuilder(RuntimeEnvironment.application, VivDatabase::class.java)
+        vivDatabase = Room.inMemoryDatabaseBuilder(RuntimeEnvironment.application, cat.xlagunas.core.data.db.VivDatabase::class.java)
             .allowMainThreadQueries()
             .build()
 
-        userEntity = UserEntity(1, "username", "firstname", "lastname", "email", "url")
+        userEntity = cat.xlagunas.core.data.db.UserEntity(1, "username", "firstname", "lastname", "email", "url")
 
         friendDao = vivDatabase.friendDao()
         userDao = vivDatabase.userDao()
@@ -110,13 +101,13 @@ class LocalContactDataSourceTest {
     @Test
     fun givenNonExistingContact_whenAdded_thenRelationshipMatchesPending() {
         val oldFriendStatus = generateFriend()
-        assert(oldFriendStatus.relationshipStatus == Relationship.NONE.name)
+        assert(oldFriendStatus.relationshipStatus == cat.xlagunas.core.data.net.Relationship.NONE.name)
 
         localContactDataSource.requestFriendship(oldFriendStatus).test()
 
         val contact = localContactDataSource.getContacts().blockingFirst()
 
-        assertThat(contact.first().relationshipStatus).isEqualToIgnoringCase(Relationship.REQUESTED.name)
+        assertThat(contact.first().relationshipStatus).isEqualToIgnoringCase(cat.xlagunas.core.data.net.Relationship.REQUESTED.name)
     }
 
     private fun generateFriend() = Friend(10, "aContact", "aName", "anImage", "anEmail", "NONE")
